创建预约

+++ lib/components/personal/appointment/data/appoint_data.dart
@@ -38,7 +38,7 @@ class AppointData extends Object{
   String endTime;
 
   @JsonKey(name: 'fee')
-  int fee;
+  double fee;
 
   @JsonKey(name: 'note')
   String note;
